app/requirements.txt
blinker==1.8.2
click==8.1.7
DateTime==5.5
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pytz==2024.1
Werkzeug==3.0.3
zope.interface==6.4.post2
